+++ crates/installer/src/downloader/core/error.rs
