+++ crates/installer/src/downloader/core/progress.rs
